Fix jshell eval exception
