+++ build.gradle
@@ -45,6 +45,8 @@ dependencies {
     // Command-line arguments
     compile 'info.picocli:picocli:4.3.2'
     annotationProcessor 'info.picocli:picocli-codegen:4.3.2'
+    // Jshell for eval
+    compile 'ch.obermuhlner:jshell-scriptengine:1.1.0'
     // Testing
     testImplementation 'org.junit.jupiter:junit-jupiter:5.6.2'
     testImplementation "org.assertj:assertj-core:3.16.1"
@@ -52,8 +54,6 @@ dependencies {
     compile group: 'org.apache.logging.log4j', name: 'log4j-core', version: '2.8.2'
     compile group: 'org.slf4j', name: 'slf4j-api', version: '1.7.25'
     compile group: 'org.apache.logging.log4j', name: 'log4j-slf4j-impl', version: '2.8.2'
-    //jshell
-    compile 'ch.obermuhlner:jshell-scriptengine:1.1.0'
 }
 
 group = 'com.tisawesomeness'
