+++ src/main/java/com/tisawesomeness/minecord/command/admin/EvalCommand.java
@@ -4,6 +4,7 @@ import com.tisawesomeness.minecord.command.CommandContext;
 import com.tisawesomeness.minecord.command.Result;
 import com.tisawesomeness.minecord.config.serial.BotListConfig;
 import com.tisawesomeness.minecord.config.serial.Config;
+import com.tisawesomeness.minecord.util.EvalUtil;
 import com.tisawesomeness.minecord.util.MessageUtils;
 
 import lombok.NonNull;
@@ -17,23 +18,17 @@ import net.dv8tion.jda.api.utils.MarkdownUtil;
 import javax.script.ScriptEngine;
 import javax.script.ScriptEngineManager;
 import javax.script.ScriptException;
-import java.lang.reflect.Field;
-import java.lang.reflect.Method;
-import java.lang.reflect.Modifier;
-import java.lang.reflect.Type;
 import java.time.OffsetDateTime;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 public class EvalCommand extends AbstractAdminCommand {
 
+    private static final int CODEBLOCK_LENGTH = 12;
+
     private final Map<String, Object> storage = new ConcurrentHashMap<>();
+    private final EvalUtil h = new EvalUtil(); // ScriptEngine needs a reference to an object
 
     public @NonNull String getId() {
         return "eval";
@@ -65,8 +60,7 @@ public class EvalCommand extends AbstractAdminCommand {
             engine.put("member", e.getMember());
         }
         engine.put("storage", storage);
-        Function<Object, String> help = EvalCommand::help;
-        engine.put("help", help);
+        engine.put("h", h);
 
         // Extract code from message
         String code = ctx.joinArgs();
@@ -77,6 +71,7 @@ public class EvalCommand extends AbstractAdminCommand {
         try {
             output = engine.eval(code);
         } catch (ScriptException ex) {
+            ex.printStackTrace();
             exMsg = ex.getMessage() == null ? "Null Script Exception" : clean(ex.getMessage(), ctx.getConfig());
         }
         if (output == null) {
@@ -86,12 +81,12 @@ public class EvalCommand extends AbstractAdminCommand {
         // Build embed
         EmbedBuilder eb = new EmbedBuilder();
         String in = clean(code, ctx.getConfig());
-        if (in.length() > MessageEmbed.TEXT_MAX_LENGTH - 10) {
+        if (in.length() > MessageEmbed.TEXT_MAX_LENGTH - CODEBLOCK_LENGTH) {
             eb.addField("Input", "Input too long!", false);
-        } else if (in.length() > MessageEmbed.VALUE_MAX_LENGTH - 10) {
-            eb.setDescription(MarkdownUtil.codeblock("js", in));
+        } else if (in.length() > MessageEmbed.VALUE_MAX_LENGTH - CODEBLOCK_LENGTH) {
+            eb.setDescription(MarkdownUtil.codeblock("java", in));
         } else {
-            eb.addField("Input", MarkdownUtil.codeblock("js", in), false);
+            eb.addField("Input", MarkdownUtil.codeblock("java", in), false);
         }
         eb.setTimestamp(OffsetDateTime.now());
         User u = e.getAuthor();
@@ -106,19 +101,19 @@ public class EvalCommand extends AbstractAdminCommand {
 
         // Check for length
         String out = clean(output.toString(), ctx.getConfig());
-        if (out.length() > MessageEmbed.VALUE_MAX_LENGTH - 10) {
-      // Send up to 10 2000-char messages
-      List<String> lines = MessageUtils.splitLinesByLength(out, Message.MAX_CONTENT_LENGTH - 10);
+        if (out.length() > MessageEmbed.VALUE_MAX_LENGTH - CODEBLOCK_LENGTH) {
+            // Send up to 10 2000-char messages
+            List<String> lines = MessageUtils.splitLinesByLength(out, Message.MAX_CONTENT_LENGTH - CODEBLOCK_LENGTH);
             int i = 0;
-            while (i < 10 && i < lines.size()) {
-                e.getChannel().sendMessage(MarkdownUtil.codeblock("js", lines.get(i))).queue();
+            while (i < CODEBLOCK_LENGTH && i < lines.size()) {
+                e.getChannel().sendMessage(MarkdownUtil.codeblock("java", lines.get(i))).queue();
                 i++;
             }
             // Let user know if the code went over 10 messages
-            String outputMsg = i == 10 ? "Too long! Output truncated." : "See above.";
+            String outputMsg = i == CODEBLOCK_LENGTH ? "Too long! Output truncated." : "See above.";
             eb.addField("Output", outputMsg, false);
         } else {
-            eb.addField("Output", MarkdownUtil.codeblock("js", out), false);
+            eb.addField("Output", MarkdownUtil.codeblock("java", out), false);
         }
 
         return ctx.replyRaw(eb);
@@ -126,7 +121,7 @@ public class EvalCommand extends AbstractAdminCommand {
     }
 
     /**
-     * Removes all blacklisted strings and everyone/here mentions from the input.
+     * Removes all blacklisted strings and everyone/here mentions from the input as a safety measure.
      * <b>NOT GUARENTEED TO WORK IN ALL CASES. NEVER REQUEST THE BOT TOKEN OR PRINT ALL JDA OR CONFIG VALUES.</b>
      * @param s The input string
      * @return A cleaned string with blacklisted strings replaced with [redacted]
@@ -146,66 +141,6 @@ public class EvalCommand extends AbstractAdminCommand {
             }
         }
         return s.replace("@everyone", "[everyone]").replace("@here", "[here]");
-    }
-
-    private static class Test {
-        public final ArrayList<String> list = new ArrayList<>();
-    }
-
-    /**
-     * Generates a list of fields and methods for an object
-     */
-    private static String help(Object o) {
-        Class<?> clazz = o.getClass();
-        String fields = "NONE";
-        if (clazz.getFields().length > 0) {
-            fields = Arrays.stream(clazz.getFields())
-                    .sorted(Comparator.comparing(Field::getName))
-                    .map(EvalCommand::getDeclaration)
-                    .collect(Collectors.joining("\n"));
-        }
-        String methods = "NONE";
-        if (clazz.getMethods().length > 0) {
-            methods = Arrays.stream(clazz.getMethods())
-                    .filter(m -> o.getClass().equals(Object.class) || !m.getDeclaringClass().equals(Object.class))
-                    .sorted(Comparator.comparing(Method::getName))
-                    .map(EvalCommand::getSignature)
-                    .collect(Collectors.joining("\n"));
-        }
-        return String.format("%s\n\nFields:\n%s\n\nMethods:\n%s", clazz.getName(), fields, methods);
-    }
-    private static String getDeclaration(Field f) {
-        String finall = Modifier.isFinal(f.getModifiers()) ? "final " : "";
-        return finall + cleanType(f.getGenericType()) + " " + f.getName();
-    }
-    /**
-     * Generates a shortened signature for a method.
-     * Exceptions, non-static modifiers, generic parameters, and annotations are excluded for brevity.
-     * @param m The method object to generate a signature for
-     * @return The signature as a string
-     */
-    private static String getSignature(Method m) {
-        String params = Arrays.stream(m.getParameters())
-                .map(p -> cleanType(p.getType()))
-                .collect(Collectors.joining(", ")); // Comma-separated args like in "add(int x, int y)"
-        String staticc = Modifier.isStatic(m.getModifiers()) ? "static " : ""; // Only static is included for brevity
-        return String.format("%s%s %s(%s)", staticc, cleanType(m.getGenericReturnType()), m.getName(), params);
-    }
-    /**
-     * Generates a clean string for a type, transforming "java.lang.String" into "String".
-     * Takes generics into account and parses them into the diamond operator format.
-     * @param t The type reflection object
-     * @return The type name as a string
-     */
-    private static String cleanType(Type t) {
-        String typeName = t.getTypeName();
-        if (typeName.contains("<")) {
-            String[] split = typeName.split("<");
-            String type = split[0].substring(split[0].lastIndexOf('.') + 1); // Thanks -1 on failure for making this super clean
-            String generic = split[1].substring(split[1].lastIndexOf('.') + 1, split[1].length() - 1);
-            return type + "<" + generic + ">";
-        }
-        return typeName.substring(typeName.lastIndexOf('.') + 1);
     }
 
 }
