+++ src/main/resources/lang.properties
@@ -42,17 +42,17 @@ command.admin.demote.usage=<user>
 command.admin.demote.aliases=deelevate,normie,badboi
 
 command.admin.eval.name=eval
-command.admin.eval.description=Evaluates some js code.
-command.admin.eval.usage=<js code>
+command.admin.eval.description=Evaluates some java code.
+command.admin.eval.usage=<java code>
 command.admin.eval.aliases=exe,exec
 command.admin.eval.help=\
-  Evaluates some js code using the Rhino engine.\n\
+  Evaluates some java code using Jshell.\n\
   Variables: `ctx`, `bot`, `jda`, `sm`, `config`, `lang`, `dbCache`, `event`, `user`, `channel`\n\
   Not available in DMs: `member`, `guild`\n\
-  Use `help(obj)` to list an object''s fields and methods.\n\
+  Use `h.elp(obj)` to list an object''s fields and methods.\n\
   Use `storage.put(stringKey, obj)` and `storage.get(stringKey)` to store variables for later.\n\
   \n\
-  See [the docs](https://docs.oracle.com/javase/8/docs/technotes/guides/scripting/prog_guide/javascript.html) for information on accessing Java from scripts.\n\
+  Use import statements (starting with `com.tisawesomeness.minecord`) to access other Java classes.\n\
   Sensitive info such as the bot token are cleaned from the input and output. In case this fails, __**never request the bot token and never print all values of the jda or config.**__
 
 command.admin.msg.name=msg
